chore : HeaderB 분리
